backend/pyproject.toml
[project]
name = "django-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.5",
    "django-behaviors>=0.5.1",
    "django-environ>=0.12.0",
    "djangorestframework>=3.16.1",
    "django-debug-toolbar>=6.0.0",
    "django-debug-toolbar-force>=0.2",
    "pillow>=11.3.0",
    "psycopg[binary,pool]>=3.2.10",
    "django-redis>=6.0.0",
    "django-cachalot>=2.8.0",
    "uvicorn>=0.35.0",
    "whitenoise>=6.9.0",
    "django-cors-headers>=4.9.0",
    "django-filter>=25.1",
    "drf-spectacular[sidecar]>=0.28.0",
    "djangorestframework-simplejwt>=5.5.1",
    "pyjwt>=2.10.1",
    "redis>=6.4.0",
    "celery>=5.5.3",
    "django-celery-beat>=2.8.1",
    "stripe>=13.1.1",
    "django-import-export>=4.3.14",
    "boto3>=1.40.74",
    "django-storages[s3]>=1.14.6",
    "flower>=2.0.1",
    "python-slugify>=8.0.0",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "django-stubs[compatible-mypy]>=5.2.7",
    "djangorestframework-stubs[compatible-mypy]>=3.16.5",
    "flake8>=7.3.0",
    "flake8-bandit>=4.1.1",
    "flake8-django>=1.1.1",
    "flake8-isort>=6.1.2",
    "flake8-pyproject>=1.2.3",
    "flake8-walrus>=1.2.0",
    "isort>=6.0.1",
    "mixer>=7.2.2",
    "mypy>=1.17.1",
    "pre-commit>=4.3.0",
    "pytest>=8.4.1",
    "pytest-cov>=7.0.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.15.1",
    "schemathesis>=4.3.6",
    "types-requests>=2.32.4.20250913",
]

[tool.flake8]
extend-select = ["I",]
ignore = ["E501", "W503", "DJ01", "S308", "S703", "F841"]
exclude = [
    "test_*.py",
    "*test_*.py",
    "conftest.py",
    "*/tests/",
    "*/test/",
    "*/migrations/*",
    ".git/",
]

[tool.isort]
profile = "black"
line_length = 88
